+++ src/components/CaptureMatching/CandidateImages.js
@@ -19,6 +19,8 @@ import theme from '../common/theme';
 import { getDateStringLocale } from 'common/locale';
 import { getDistance } from 'geolib';
 import OptimizedImage from 'components/OptimizedImage';
+import { CopyButton } from '../common/CopyButton';
+import CopyNotification from '../common/CopyNotification';
 
 const useStyles = makeStyles({
   containerBox: {
@@ -58,6 +60,7 @@ const useStyles = makeStyles({
     display: 'flex',
     gap: theme.spacing(2),
   },
+  candidateImageNotes: { fontStyle: 'italic', paddingTop: theme.spacing(1) },
   button: {
     fontSize: '16px',
   },
@@ -84,7 +87,7 @@ const useStyles = makeStyles({
     textAlign: 'center',
   },
   candidateCaptureContainer: {
-    position: 'relative',
+    flexBasis: '250px',
   },
   captureInfo: {
     position: 'absolute',
@@ -138,6 +141,8 @@ function CandidateImages({ capture, candidateImgData, sameTreeHandler }) {
   const classes = useStyles();
 
   const [showBox, setShowBox] = useState([]);
+  const [snackbarOpen, setSnackbarOpen] = useState(false);
+  const [snackbarLabel, setSnackbarLabel] = useState('');
 
   useEffect(() => {
     const initialCandidateData = candidateImgData.map((tree) => tree.id);
@@ -151,6 +156,12 @@ function CandidateImages({ capture, candidateImgData, sameTreeHandler }) {
 
   const showImgBox = (i) => {
     setShowBox([...showBox, i]);
+  };
+
+  const confirmCopy = (label) => {
+    setSnackbarOpen(false);
+    setSnackbarLabel(label);
+    setSnackbarOpen(true);
   };
 
   return (
@@ -180,6 +191,11 @@ function CandidateImages({ capture, candidateImgData, sameTreeHandler }) {
                     <Tooltip title={tree.id}>
                       <Typography variant="h5">
                         Tree {(tree.id + '').substring(0, 10) + '...'}
+                        <CopyButton
+                          label={tree.id}
+                          value={tree.id}
+                          confirmCopy={confirmCopy}
+                        />
                       </Typography>
                     </Tooltip>
                   </Box>
@@ -215,64 +231,74 @@ function CandidateImages({ capture, candidateImgData, sameTreeHandler }) {
                 }}
               >
                 <Box className={classes.gridList} cols={3}>
-                  {(tree.captures.length ? tree.captures : [tree]).map(
+                  {(tree.captures?.length ? tree.captures : [tree]).map(
                     (candidateCapture) => {
                       return (
                         <Box
                           key={`${tree.id}_${candidateCapture.id}`}
                           className={classes.candidateCaptureContainer}
                         >
-                          <OptimizedImage
-                            src={candidateCapture.image_url}
-                            alt={`Candidate capture ${candidateCapture.id}`}
-                            objectFit="cover"
-                            width={250}
-                            style={{
-                              width: '250px',
-                              height: '100%',
-                            }}
-                            alertHeight="300px"
-                            alertTextSize=".9rem"
-                            alertTitleSize="1.2rem"
-                          />
-                          <Box className={classes.captureInfo}>
-                            <Box className={classes.captureInfoDetail}>
-                              <AccessTimeIcon />
-                              <Typography variant="body1">
-                                {getDateStringLocale(
-                                  new Date(candidateCapture.created_at)
-                                )}
-                              </Typography>
-                            </Box>
-                            <Box className={classes.captureInfoDetail}>
-                              <LocationOnOutlinedIcon
-                                style={{
-                                  cursor: 'pointer',
-                                }}
-                                onClick={() => {
-                                  window.open(
-                                    `https://www.google.com/maps/search/?api=1&query=${capture.latitude},${capture.longitude}`
-                                  );
-                                }}
-                              />
-                              <Typography variant="body1">
-                                {capture?.latitude && capture?.longitude && (
-                                  <DistanceTo
-                                    lat1={capture.latitude}
-                                    lon1={capture.longitude}
-                                    lat2={
-                                      candidateCapture.latitude ||
-                                      candidateCapture.lat
-                                    }
-                                    lon2={
-                                      candidateCapture.longitude ||
-                                      candidateCapture.lon
-                                    }
-                                  />
-                                )}
-                              </Typography>
+                          <Box style={{ position: 'relative' }}>
+                            <OptimizedImage
+                              src={candidateCapture.image_url}
+                              alt={`Candidate capture ${candidateCapture.id}`}
+                              objectFit="cover"
+                              width={250}
+                              style={{
+                                width: '250px',
+                                height: '100%',
+                              }}
+                              alertHeight="300px"
+                              alertTextSize=".9rem"
+                              alertTitleSize="1.2rem"
+                            />
+                            <Box className={classes.captureInfo}>
+                              <Box className={classes.captureInfoDetail}>
+                                <AccessTimeIcon />
+                                <Typography variant="body1">
+                                  {getDateStringLocale(
+                                    new Date(
+                                      candidateCapture.captured_at ||
+                                      candidateCapture.created_at
+                                    )
+                                  )}
+                                </Typography>
+                              </Box>
+                              <Box className={classes.captureInfoDetail}>
+                                <LocationOnOutlinedIcon
+                                  style={{
+                                    cursor: 'pointer',
+                                  }}
+                                  onClick={() => {
+                                    window.open(
+                                      `https://www.google.com/maps/search/?api=1&query=${capture.latitude},${capture.longitude}`
+                                    );
+                                  }}
+                                />
+                                <Typography variant="body1">
+                                  {capture?.latitude && capture?.longitude && (
+                                    <DistanceTo
+                                      lat1={capture.latitude}
+                                      lon1={capture.longitude}
+                                      lat2={
+                                        candidateCapture.latitude ||
+                                        candidateCapture.lat
+                                      }
+                                      lon2={
+                                        candidateCapture.longitude ||
+                                        candidateCapture.lon
+                                      }
+                                    />
+                                  )}
+                                </Typography>
+                              </Box>
                             </Box>
                           </Box>
+                          {candidateCapture.notes && (
+                            <Typography className={classes.candidateImageNotes}>
+                              {candidateCapture.notes}
+                            </Typography>
+                          )}
                         </Box>
                       );
                     }
@@ -306,6 +332,11 @@ function CandidateImages({ capture, candidateImgData, sameTreeHandler }) {
             </Paper>
           );
         })}
+      <CopyNotification
+        snackbarLabel={snackbarLabel}
+        snackbarOpen={snackbarOpen}
+        setSnackbarOpen={setSnackbarOpen}
+      />
     </Box>
   );
 }