+++ src/components/Captures/CaptureTable.test.js
@@ -84,11 +84,11 @@ describe('Captures', () => {
       expect(arr[1]).toBe('1-4 of 4');
     });
 
-    it('should have 13 headers', () => {
+    it('should have 14 headers', () => {
       const table = screen.getByRole(/table/i);
       const headers = within(table).getAllByRole(/columnheader/i);
       const arr = headers.map((header) => header.textContent);
-      expect(arr).toHaveLength(13);
+      expect(arr).toHaveLength(14);
     });
 
     it('renders headers for captures table', () => {
@@ -108,6 +108,8 @@ describe('Captures', () => {
       item = within(table).getByText(/Species/i);
       expect(item).toBeInTheDocument();
       item = within(table).getByText(/Token Id/i);
+      expect(item).toBeInTheDocument();
+      item = within(table).getByText(/Wallet/i);
       expect(item).toBeInTheDocument();
       item = within(table).getByText(/Capture Tags/i);
       expect(item).toBeInTheDocument();