+++ src/components/EditGrower.js
@@ -10,20 +10,23 @@ import {
   TextField,
   CircularProgress,
 } from '@material-ui/core';
-import api from '../api/growers';
 import ImageScroller from './ImageScroller';
 import SelectOrg from './common/SelectOrg';
-import { GrowerContext } from '../context/GrowerContext';
-import { ORGANIZATION_NOT_SET } from '../models/Filter';
+import { GrowerContext } from 'context/GrowerContext';
+import { ORGANIZATION_NOT_SET } from 'models/Filter';
 
 const useStyle = makeStyles((theme) => ({
   container: {
-    position: 'relative',
     padding: theme.spacing(0, 4),
+  },
+  textContainer: {
+    display: 'flex',
+    flexFlow: 'row nowrap',
+    justifyContent: 'space-between',
   },
   textInput: {
     margin: theme.spacing(2, 0),
-    flexGrow: 1,
+    flex: '0 0 49%',
   },
 }));
 
@@ -40,7 +43,7 @@ const EditGrower = (props) => {
     async function loadGrowerImages() {
       if (grower?.id) {
         setLoadingGrowerImages(true);
-        const selfies = await api.getGrowerSelfies(grower.id);
+        const selfies = await growerContext.getGrowerSelfies(grower.id);
         setLoadingGrowerImages(false);
 
         setGrowerImages([
@@ -54,27 +57,11 @@ const EditGrower = (props) => {
     loadGrowerImages();
   }, [grower]);
 
-  async function updateGrower(grower) {
-    await api.updateGrower(grower);
-    const updatedGrower = await api.getGrower(grower.id);
-    // only update context if growers have already been downloaded
-    if (growerContext.growers.length) {
-      const index = growerContext.growers.findIndex(
-        (p) => p.id === updatedGrower.id
-      );
-      if (index >= 0) {
-        const growers = [...growerContext.growers];
-        growers[index] = updatedGrower;
-        growerContext.updateGrowers(growers);
-      }
-    }
-  }
-
   async function handleSave() {
     if (growerUpdate) {
       setSaveInProgress(true);
       // TODO handle errors
-      await updateGrower({
+      await growerContext.updateGrower({
         id: grower.id,
         ...growerUpdate,
       });
@@ -84,6 +71,7 @@ const EditGrower = (props) => {
   }
 
   function handleCancel() {
+    setGrowerUpdate(null);
     onClose();
   }
 
@@ -95,21 +83,12 @@ const EditGrower = (props) => {
       return newGrower[key] !== grower[key];
     });
 
-    if (changed) {
-      setGrowerUpdate(newGrower);
-    } else {
-      setGrowerUpdate(null);
-    }
+    changed ? setGrowerUpdate(newGrower) : setGrowerUpdate(null);
   }
 
   function getValue(attr) {
     // Ensure empty strings are not overlooked
-    if (growerUpdate?.[attr] != null) {
-      return growerUpdate[attr];
-    } else if (grower[attr] != null) {
-      return grower[attr];
-    }
-    return '';
+    return growerUpdate?.[attr] ?? grower?.[attr] ?? '';
   }
 
   const inputs = [