+++ src/context/TagsContext.js
@@ -1,6 +1,7 @@
-import React, { useState, createContext, useEffect } from 'react';
+import React, { useContext, useState, createContext, useEffect } from 'react';
 import api from '../api/treeTrackerApi';
-import { getOrganizationUUID } from '../api/apiUtils';
+import { AppContext } from './AppContext';
+
 import * as loglevel from 'loglevel';
 
 const log = loglevel.getLogger('../context/TagsContext');
@@ -14,31 +15,33 @@ export const TagsContext = createContext({
 });
 
 export function TagsProvider(props) {
+  const { getOrganizationUUID } = useContext(AppContext);
   const [tagList, setTagList] = useState([]);
   const [tagInput, setTagInput] = useState([]);
+  const orgId = getOrganizationUUID();
 
   useEffect(() => {
     const abortController = new AbortController();
-    loadTags({ signal: abortController.signal });
+    loadTags(orgId, { signal: abortController.signal });
     return () => abortController.abort();
-  }, []);
+  }, [orgId]);
 
   // EVENT HANDLERS
-  const loadTags = async () => {
-    const response = await api.getTags();
+  const loadTags = async (orgId, abortSignal) => {
+    const response = await api.getTags(orgId, abortSignal);
     setTagList(response.tags);
   };
+
   /*
    * check for new tags in tagInput and add them to the database
    */
-
-  const createTags = async () => {
+  const createTags = async (newTag) => {
     const orgId = getOrganizationUUID();
     const newTagTemplate = {
       isPublic: orgId ? false : true,
       owner_id: orgId,
     };
-    const promises = tagInput.map(async (t) => {
+    const promises = [...tagInput, newTag].map(async (t) => {
       const existingTag = tagList.find((tag) => tag.name === t);
       if (!existingTag) {
         return api.createTag({ ...newTagTemplate, name: t });